First commit for toggling mentions using shortcut and button
Also did some refactoring and moved code to actions

Fixed coding style errors
